+++ lib/Effects/Effects/ColorStrobe.h
@@ -6,21 +6,21 @@ void colorStrobeEffectLoop(struct EffectState *effectState)
   const CRGB RED = CRGB(255, 0, 0);
   const CRGB BLUE = CRGB(255, 0, 0);
 
-  const unsigned int blinkTs = 20;
-  const unsigned int steps = 10;
-  const unsigned int loopMs = 1000;
+  const uint8_t blinkTs = 20;
+  const uint8_t steps = 10;
+  const uint8_t loopMs = 1000;
 
-  const unsigned int loopPosition = effectState->loopPosition % loopMs;
+  const uint8_t loopPosition = effectState->loopPosition % loopMs;
   const float loopPercent = float(loopPosition) / loopMs;
-  const unsigned int step = loopPercent * steps;
-  const unsigned int stepBase = (loopMs / steps) * step;
+  const uint8_t step = loopPercent * steps;
+  const uint8_t stepBase = (loopMs / steps) * step;
 
-  const unsigned int hueLow = 170;
-  const unsigned int hueHigh = 300;
+  const uint8_t hueLow = 170;
+  const uint8_t hueHigh = 300;
 
-  const unsigned int hue = hueLow + (hueHigh - hueLow) * loopPercent;
+  const uint8_t hue = hueLow + (hueHigh - hueLow) * loopPercent;
 
-  for (int i = 0; i < config.LED_CNT; i++)
+  for (uint16_t i = 0; i < config.LED_CNT; i++)
   {
     if (loopPosition < stepBase + blinkTs)
     {
