+++ lib/Effects/Effects/Throb.h
@@ -2,12 +2,12 @@
 
 void throbEffectLoop(struct EffectState *effectState, int hue)
 {
-  const unsigned int loopMs = 3000;
-  const unsigned int loopPosition = effectState->loopPosition % loopMs;
+  const uint8_t loopMs = 3000;
+  const uint8_t loopPosition = effectState->loopPosition % loopMs;
   const float loopPercent = float(loopPosition) / loopMs;
-  const unsigned int intensity = abs(cos(PI * .5 * loopPercent * 2) * 255);
+  const uint8_t intensity = abs(cos(PI * .5 * loopPercent * 2) * 255);
 
-  for (int i = 0; i < config.LED_CNT; i++)
+  for (uint16_t i = 0; i < config.LED_CNT; i++)
   {
     leds[i] = CHSV(hue, 255, intensity);
   }
